Add dummy articles and pass props to article list components
- Introduce dummy articles in the home screen index
- Update FeaturedArticleList and ArticleFlatList to accept article props
- Remove hardcoded dummy articles from individual components
- Use key prop with article id for list rendering
